Pull up fetching the teams for components in a separate function
